refactor: scope feature services and stores
