+++ app/frontend/src/features/generation/composables/useGenerationTransport.ts
@@ -2,7 +2,7 @@ import {
   extractGenerationErrorMessage,
   type GenerationQueueClient,
   type GenerationWebSocketManager,
-} from '@/services';
+} from '@/features/generation/services';
 import type {
   GenerationCompleteMessage,
   GenerationErrorMessage,
