+++ app/frontend/src/features/history/composables/useGenerationHistory.ts
@@ -2,7 +2,8 @@ import { computed, ref, type ComputedRef } from 'vue';
 import type { MaybeRefOrGetter } from 'vue';
 
 import { debounce, type DebouncedFunction } from '@/utils/async';
-import { listResults as listHistoryResults, useBackendClient } from '@/services';
+import { useBackendClient } from '@/services';
+import { listResults as listHistoryResults } from '@/features/history/services';
 import type {
   GenerationHistoryQuery,
   GenerationHistoryResult,
