+++ app/frontend/src/services/index.ts
@@ -1,7 +1,4 @@
 export * from './apiClient';
 export * from './backendClient';
-export * from '../features/generation/services';
 export * from './system';
 export * from './lora';
-export * from '../features/history/services';
-export * from '../features/analytics/services';
