+++ app/frontend/src/stores/index.ts
@@ -1,7 +1,5 @@
 export * from './settings';
 export * from './app';
 export * from './adminMetrics';
-export * from '../features/generation/stores';
 export * from './adapterCatalog';
-export * from '../features/analytics/stores';
