+++ app/frontend/src/utils/promptGeneration.ts
@@ -1,4 +1,4 @@
-import { createGenerationParams, requestGeneration } from '@/services';
+import { createGenerationParams, requestGeneration } from '@/features/generation/services';
 
 import type { CompositionEntry } from '@/types';
